+++ src/__tests__/pages/teachers/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug]/share.test.tsx
@@ -322,13 +322,16 @@ describe("pages/teachers/lessons/[lessonSlug]/downloads", () => {
         ogSiteName: "NEXT_PUBLIC_SEO_APP_NAME",
         title:
           "Lesson Share: Islamic Geometry | KS4 Maths | NEXT_PUBLIC_SEO_APP_NAME",
-        description: "Lesson share",
+        description:
+          "Share online lesson activities with your students, such as videos, worksheets and quizzes.",
         ogTitle:
           "Lesson Share: Islamic Geometry | KS4 Maths | NEXT_PUBLIC_SEO_APP_NAME",
-        ogDescription: "Lesson share",
+        ogDescription:
+          "Share online lesson activities with your students, such as videos, worksheets and quizzes.",
         ogUrl: "NEXT_PUBLIC_SEO_APP_URL",
-        canonical: "NEXT_PUBLIC_SEO_APP_URL",
-        robots: "noindex,nofollow",
+        canonical:
+          "NEXT_PUBLIC_SEO_APP_URL/teachers/programmes/maths-higher-ks4-l/units/geometry/lessons/macbeth-lesson-1",
+        robots: "index,follow",
       });
     });
   });